@@ -11,63 +11,45 @@
     "deploy": "python src/deployment/bucket.py"
   },
   "author": "Joel Roxell <joel.roxell@annevo.com> (https://github.com/JoelRoxell)",
-  "devDependencies": {
-    "@types/aws-sdk": "^2.7.0",
+  "dependencies": {
     "@types/axios": "^0.14.0",
     "@types/hoist-non-react-statics": "^3.3.1",
-    "@types/koa": "^2.0.48",
-    "@types/lodash": "^4.14.134",
-    "@types/react": "^16.8.22",
+    "@types/react": "^16.8.23",
     "@types/react-dom": "^16.8.4",
-    "@types/react-redux": "^7.0.1",
     "@types/react-router-dom": "^4.3.4",
-    "@types/webpack-env": "^1.13.9",
-    "autoprefixer": "^9.6.0",
-    "aws-sdk": "^2.481.0",
+    "@types/webpack-env": "^1.14.0",
+    "autoprefixer": "^9.6.1",
     "axios": "^0.19.0",
-    "css-loader": "^3.0.0",
+    "css-loader": "^3.1.0",
     "dts-css-modules-loader": "^1.0.1",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "fork-ts-checker-webpack-plugin": "^1.3.7",
+    "fork-ts-checker-webpack-plugin": "^1.4.3",
     "html-webpack-plugin": "^3.2.0",
-    "koa": "^2.7.0",
     "node-sass": "^4.12.0",
     "postcss-flexbugs-fixes": "^4.1.0",
     "postcss-loader": "^3.0.0",
     "prepend-style-loader": "^2.0.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
-    "react-redux": "^7.0.1",
     "react-router-dom": "^5.0.1",
-    "redux": "^4.0.1",
-    "redux-devtools-extension": "^2.13.8",
-    "redux-thunk": "^2.3.0",
     "rimraf": "^2.6.3",
     "sass-loader": "^7.1.0",
     "style-loader": "^0.23.1",
     "ts-loader": "^6.0.4",
     "tslint": "^5.18.0",
-    "typescript": "^3.5.2",
+    "typescript": "^3.5.3",
     "typings-for-css-modules-loader": "^1.7.0",
-    "webpack": "^4.35.0",
-    "webpack-cli": "^3.3.5",
-    "webpack-dev-server": "^3.7.2"
-  },
-  "dependencies": {
-    "@types/classnames": "^2.2.8",
-    "@types/moment": "^2.13.0",
-    "@types/react-dates": "^17.1.5",
-    "@types/react-stripe-elements": "^1.1.10",
-    "@types/webpack": "^4.4.34",
+    "webpack": "^4.36.1",
+    "webpack-cli": "^3.3.6",
+    "webpack-dev-server": "^3.7.2",
+    "@types/classnames": "^2.2.9",
+    "@types/webpack": "^4.32.0",
     "classnames": "^2.2.6",
-    "file-loader": "^4.0.0",
+    "file-loader": "^4.1.0",
     "generator-annevo": "^1.1.6-beta.2",
-    "moment": "^2.24.0",
+    "net": "^1.0.2",
     "react-addons-css-transition-group": "^15.6.2",
-    "react-dates": "^20.2.4",
-    "react-ionicons": "^3.0.1",
-    "react-stripe-elements": "^3.0.0",
-    "url-loader": "^2.0.0"
+    "url-loader": "^2.1.0"
   },
   "browserslist": [
     ">1%",
