@@ -1,24 +1,17 @@
 import * as React from 'react'
-import { connect } from 'react-redux'
 
-import { ConnectedReduxProps } from 'store'
-import { signOut } from 'store/auth'
 import { Redirect } from 'react-router'
 
 import * as style from './SignOut.scss'
 
-class SignOut extends React.Component<ConnectedReduxProps> {
-  componentWillMount() {
-    this.props.dispatch(signOut())
-  }
+function SignOut() {
+  React.useEffect(() => console.log('Sign out'), [])
 
-  render() {
-    return (
-      <div className={style.signOut}>
-        <Redirect to="/" />
-      </div>
-    )
-  }
+  return (
+    <div className={style.signOut}>
+      <Redirect to="/" />
+    </div>
+  )
 }
 
-export default connect()(SignOut)
+export default SignOut
